create core.random
