+++ sympy/combinatorics/tests/test_schur_number.py
@@ -1,6 +1,6 @@
 from sympy.core import S, Rational
 from sympy.combinatorics.schur_number import schur_partition, SchurNumber
-from sympy.testing.randtest import _randint
+from sympy.core.random import _randint
 from sympy.testing.pytest import raises
 from sympy.core.symbol import symbols
