+++ sympy/functions/combinatorial/tests/test_comb_numbers.py
@@ -492,7 +492,7 @@ def test_nC_nP_nT():
         nP, nC, nT, stirling, _stirling1, _stirling2, _multiset_histogram, _AOP_product)
 
     from sympy.combinatorics.permutations import Permutation
-    from random import choice
+    from sympy.core.random import choice
 
     c = string.ascii_lowercase
     for i in range(100):
