+++ sympy/functions/elementary/tests/test_miscellaneous.py
@@ -396,7 +396,7 @@ def test_issue_11099():
     assert Max(x, y).evalf(subs=fixed_test_data) == \
         Max(x, y).subs(fixed_test_data).evalf()
     # randomly generate some test data
-    from random import randint
+    from sympy.core.random import randint
     for i in range(20):
         random_test_data = {x: randint(-100, 100), y: randint(-100, 100)}
         assert Min(x, y).evalf(subs=random_test_data) == \
