+++ sympy/functions/special/tests/test_bessel.py
@@ -18,7 +18,7 @@ from sympy.series.order import O
 from sympy.series.series import series
 from sympy.functions.special.bessel import (airyai, airybi,
                                             airyaiprime, airybiprime, marcumq)
-from sympy.testing.randtest import (random_complex_number as randcplx,
+from sympy.core.random import (random_complex_number as randcplx,
                                       verify_numerically as tn,
                                       test_derivative_numerically as td,
                                       _randint)
@@ -483,7 +483,7 @@ def test_branching():
     assert besseli(n, polar_lift(x)) == besseli(n, x)
 
     def tn(func, s):
-        from random import uniform
+        from sympy.core.random import uniform
         c = uniform(1, 5)
         expr = func(s, c*exp_polar(I*pi)) - func(s, c*exp_polar(-I*pi))
         eps = 1e-15
