+++ sympy/integrals/tests/test_integrals.py
@@ -39,7 +39,7 @@ from sympy.integrals.risch import NonElementaryIntegral
 from sympy.physics import units
 from sympy.testing.pytest import (raises, slow, skip, ON_TRAVIS,
     warns_deprecated_sympy)
-from sympy.testing.randtest import verify_numerically
+from sympy.core.random import verify_numerically
 
 
 x, y, a, t, x_1, x_2, z, s, b = symbols('x y a t x_1 x_2 z s b')
