+++ sympy/ntheory/tests/test_egyptian_fraction.py
@@ -2,7 +2,7 @@ from sympy.core.numbers import Rational
 from sympy.ntheory.egyptian_fraction import egyptian_fraction
 from sympy.core.add import Add
 from sympy.testing.pytest import raises
-from sympy.testing.randtest import random_complex_number
+from sympy.core.random import random_complex_number
 
 
 def test_egyptian_fraction():
