+++ sympy/polys/factortools.py
@@ -977,7 +977,7 @@ def dmp_zz_wang(f, u, K, mod=None, seed=None):
     .. [2] [Geddes92]_
 
     """
-    from sympy.testing.randtest import _randint
+    from sympy.core.random import _randint
 
     randint = _randint(seed)
