+++ sympy/solvers/tests/test_solveset.py
@@ -32,7 +32,7 @@ from sympy.tensor.indexed import Indexed
 from sympy.utilities.iterables import numbered_symbols
 
 from sympy.testing.pytest import (XFAIL, raises, skip, slow, SKIP, _both_exp_pow)
-from sympy.testing.randtest import verify_numerically as tn
+from sympy.core.random import verify_numerically as tn
 from sympy.physics.units import cm
 
 from sympy.solvers import solve
