+++ sympy/utilities/randtest.py
@@ -2,8 +2,8 @@ from sympy.utilities.exceptions import SymPyDeprecationWarning
 
 SymPyDeprecationWarning(
     feature="Import sympy.utilities.randtest",
-    useinstead="Import from sympy.testing.randtest",
+    useinstead="Import from sympy.core.random",
     issue=18095,
     deprecated_since_version="1.6").warn()
 
-from sympy.testing.randtest import *  # noqa:F401
+from sympy.core.random import *  # noqa:F401
